Move to its own compilation unit
